Move layout preview into an individual class.
